A few more misc fixes for the linux port.
Temporarily turn off SSL until the SSL full duplex patch can land.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/272004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28550 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b172068c7dea600ebda1c7fdfd31dca03d10b119

+++ net/flip/flip_session.h
@@ -56,7 +56,7 @@ class PrioritizedIOBuffer {
 
   IOBuffer* buffer() const { return buffer_; }
 
-  int size() const { return buffer_->size(); }
+  size_t size() const { return buffer_->size(); }
 
   void release() { buffer_ = NULL; }
